requirements.txt
python-dotenv
fastapi[standard]

requests
PyJWT
cryptography
httpx